+++ utoipa-gen/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
 
+* Finalize actix-web utoipa bindings (https://github.com/juhaku/utoipa/pull/1160)
 *  Enhance no_recursion rule to apply also containers (https://github.com/juhaku/utoipa/pull/1144)
 
 ## 5.1.0 - Oct 16 2024